refactor(sn): reuse system message sender connections
Whenever we receive a valid system message, we now merge the sender's
connection into our network knowledge, if applicable.

An example of when this is useful is when a node uses their SAP to cold-
call an elder who they have not yet spoken to. Without this change,
there would be additional connections since the elder would reply over a
new connection. With this change in place, they will instead be able to
reply over the same sending connection.