refactor: parse different output from `nrs create`
This command now returns a `Url`, along with a tuple of 3 strings. The strings are:
* a "+" sign to indicate the addition of new content
* the NRS name that was created
* the Safe URL the entry points to

The test helper function that parses the output of the command was updated to return these types and
any callers were updated to use a `Url` directly. Previously, the caller was expecting the Url back
as a `String`, and most were then creating a `Url` from that `String`, so this cuts the
intermediate step.