refactor(sn): only recreate failed connections in `Comm::send`
`Peer`s may hold connections that have been lost due to timeouts etc. As
such, we want to handle connection loss when sending by trying again and
forcing a reconnection. We would previously do this by passing a
`force_reconnection` flag, however in the event of concurrent sends this
might lead to more connections being opened than necessary.

The new approach tests the existing connection to see if it's the same
one that previously failed, and will only reconnect if so.

This does not fully protect from 'stampedes' of new connections, since
the same issue exists when the `Peer` has no connection (e.g. many
concurrent sends for the same peer may lead to many connections being
made), but this will be followed up separately.