refactor(sn): prevent concurrent sends from opening many connections
This is an optimisation in favour of reducing the number of connections
we open, at the cost of reduced parallelism when sending multiple
messages to the same unconnected `Peer` (which is fairly common when
performing DKG rounds with new peers, e.g.).

The implementation is based on giving `Peer` the responsibility of
creating new connections if necessary, meaning it can take the write
lock and see if any competing attempt already set a new connection - and
if so, avoid creating another one. A 'fast path' is included that will
return immediately if there's an existing valid connection.