refactor(sn): prefer existing connection when merging `Peer`s
This is a probably-premature optimisation to minimise when we take write
locks on `Peer::connection`. Previously when merging connections we
would prioritise the source connection (e.g. we would overwrite an
existing connection in the target). In particular, this would cause
write-locking even if it was the same connection, which could lead to
needless contention when trying to read connections.

This commit changes to instead retain the existing connection if we have
one, and does so in a way that minimises the likelihood that we need to
take a write lock at all.