fix(sn): restore `ServiceMsg` authority check
The `AuthorityProof` struct is designed to be a proof of valid
authority, by ensuring all possible constructors either generate or
validate a signature. This can only be guaranteed if the field remains
module-private. At some point it seems the field was made `pub(crate)`,
which meant we were missing an authority check for some `ServiceMsg`s,