chore: quickly upload musl build to s3
This was added as a quick utility, to be used with the testnet tool. For some reason, when I use the
testnet tool and try to specify a local binary to upload, the Terraform provisioner doesn't work
correctly. This target just runs a build and then uploads it in an archive as the 'latest' version,
which is what the testnet tool will pull down by default.