feat(sn_cli): genesis key argument for `node join`
Provides and correctly sets the genesis key argument for the `node join` command, which was
previously using a hard-coded placeholder value.

This isn't actually just a feature commit, it's also a refactor to make the command unit-testable
using the `NetworkLauncher` trait. Unit test coverage has been provided for all the code paths
in the command. More elaborate documentation has also been provided for each of the arguments for
the command.

In addition to the unit tests, I've tested the command for joining a local and a remote network,
both of which were successful.