ci: run sn_cli unit tests for every pull request
This does involve building the code for the CLI, but that shouldn't be much of an overhead given
that we've already built the safe_network code. These tests are true unit tests, so they take less
than a second to run. I figure we may as well just run them for every PR.