chore: unify skip port forwarding arguments
We decided to rename the `--skip-igd` argument to `--skip-auto-port-forwarding`. I had initially
renamed the arguments on the CLI and `sn_launch_tool` to `--disable-port-forwarding`, but it turned
out that wasn't a completely accurate description of what's happening.

We are now uniformly referring to this argument as `--skip-auto-port-forwarding`, which is quite an
accurate description, as it skips (but not disables) the software-based port forwarding, leaving you
to 'manually' setup port forwarding on your router, if need be.

Also fixed some clippy warnings.