tests: cross platform support for cli unit tests
On Windows, some tests were failing due to issues with file paths and around URL parsing.

For the files tests, it was because strings were being compared rather than `Path` structs.

For the URL related failures, somewhat strangely, on Windows, the `parse` function on the `url`
crate returned an `Ok` result if you gave it a file path, which it doesn't do on Linux. A quick
solution was to test the URL for a host.