feat(sn_cli)!: `nrs create` only creates topnames
BREAKING CHANGE: this command should now be used to create new topnames, but *not* a subname and a
new topname at the same time. The user is advised on this by way of suggestions in error messages.

The user experience of this command was also improved in the following ways:
* Validate URLs supplied with the `--link` argument *before* creating the topname. This is to ensure
  the user doesn't receive an error message after the topname was already created, which can
  potentially be quite confusing.
* Suggest supplying valid XorUrls for the `--link` argument.
* Suggest that links to versionable content must have their version hash specified on the `--link`
  argument.
* Suggest the use of the `add` command with the `--create-top-name` argument to create a subname and
  a topname at the same time, or use the `create` command to first create the topname, then use the
  `add` command to create a subname.
* Suggest trying a different name if the chosen topname is already registered.
* The grammar of a couple of error messages in the API were updated for improved readability.

Some behaviour was removed:
* Since dry run support is still in progress for NRS, the `--dry-run` code was removed.
* The `--direct` argument was removed, since it only really applies to subnames, which you can no
  longer supply with the `create` command. This is a BREAKING CHANGE.

The initial intention was to fix up the failing NRS tests for the CLI. These failures were due
to the fact that the API for NRS had changed, which was resulting in the commands behaving a bit
differently than they used to. The main difference being the fact that the NRS API doesn't allow the
creation of subnames along with a topname.

I took the opportunity to introduce new tests that should cover all the scenarios for the `create`
command.