feat(sn_cli): refined ux and behaviour for `nrs add`
The behaviour for the command is mostly retained, but the user experience has been improved in the
following ways:

* The documentation for the command and the arguments were elaborated to provide more information.
* Text was added to the command output when the `--default` argument is used, to indicate that a
  link was created to the topname.
* Suggest that links to versionable content must have their version hash specified on the `--link`
  argument.

BREAKING CHANGES:
* The `--name` argument has been renamed to `--public-name`.
* The `--direct` flag argument was removed.

Tests were provided for all of the scenarios and code paths explored in the command and there was
some refactoring done to share some code between the `create` and `add` commands. The calls to
`nrs_add` and `nrs_associate` were wrapped in functions, due to the slightly verbose error handling
required for providing suggestions to the user.

Finally, the dry run functionality was removed since this currently isn't available for NRS. We can
add it back in when it is.