tests: add coverage for `nrs remove` command
There are no behavioural changes in the command, but there was a little bit of code refactoring and
some improved user experience by way of error messages and suggestions.

We need to come back and revisit some odd behaviour with this command, related to removing topnames
and attempting to remove subnames that don't exist.