feat(sn_cli): rename nrs `create` to `register`
BREAKING CHANGE: this command now has a different name.

We decided it was worth having increased distinction between subname and topname management, where
`register` and `unregister` will be for topname management, and `add` and `remove` will be for
subname management. Behaviourally, everything is the same.

Arguments and string references were replaced with the relevant verbs.