chore: use versioned self update crate
Currently the `self_update` crate has an issue whereby the wrong download URL is assigned to assets
on the Github release. We've fixed this issue on our own fork of the crate, and we're currently
speaking to the owner to try and get it merged in, but there could potentially be an indefinite
amount of time in getting that merged and a new version of that crate released.

In the mean time, we really need to get a release of `sn_cli`, but with our new release process, we
can't do so without having a it published on crates.io. Therefore, this commit is reverting back to
the versioned `self_update` crate that has the problem in it, but the next commit will provide some
temporary code to work around the URL problem.

As part of the work here, I've made a few changes to ensure that the CLI can actually build if the
`self-update` feature is switched off, and I also removed an optional dependency on the `reqwest`
crate, which was tied to the `self-update` feature. The reason for this is because there's some code
in one of the `networks` commands that can download a network config from a remote location, and we
need to make an HTTP request for that. I realise `reqwest` may be a large dependency, but in my
view, it doesn't make sense to tie this code to the `self-update` feature, since the `networks`
commands have no relation to `self_update`.