refactor: build asset download url in cli
The asset download URL is being incorrectly assigned in the `self_update` crate. As per the previous
commit, we're trying to resolve this upstream, but don't know how long it will take.

Instead, we use a temporary workaround, which is our own code to determine the download URL. We are
still `self_update` to download and extract the release archives. When the issue is fixed, we can
remove this code.

As part of this work, I added some integration tests for the `node install` command and tidied up
some of its documentation. These tests will run in the CLI test suite. To enable better testing, an
environment variable was added to specify the location of the CLI's config directory; we don't need
to expose this as an option to users, though it would be there if we wanted to.