ci: disable self update feature in sn_cli
Due to moving everything into one repository, the self update feature isn't working at the moment.
We will need to revisit and fix it. For now, I'm just disabling the feature from the release build
of the CLI.

Also fix a clippy warning that ideally should have been done on the last commit.