chore: remove use of `indicatif` crate
This was another crate that was using a git reference. It provides progress bars and other
mechanisms for updating terminal user interfaces with progress on long running tasks. We were using
this crate to display some progress bars during the `files get` command.

Obviously, removing the crate means we no longer have the progress bars in the UI, but we already
had a text-based mechanism for displaying progress, so for the time being, we can fall back on this.
We can look at other progress bar crates or refactor `files get` to work in parallel.

The main crate used a design that made use of threads for handling the display and update of
multiple progress bars, but our code for `files get` is synchronous. The forked crate was used
because it had a mechanism for updating the display in a single-threaded manner.

Removing this crate now means we should be able to get `sn_cli` published, which I would argue is
currently of higher priority.