chore: update cli install script for single repo
Modify the install process so that it downloads the CLI directly from the Github release rather than
an S3 bucket, which doesn't get updated any more.

Unfortunately, we lose some functionality here, namely, the ability to install a specific version of
the CLI. The reason for this is because of the way the release process is currently setup to assign
the tag name to the Github Release. Since we now have 3 crates in the same repo, the tag name that
we're currently assigning is just the version of the `safe_network` crate. So we can't query for a
specific version of the CLI. We can perhaps change the release process so that the tag name is
assigned the version number of each component, and then we'd be able to go for a specific CLI
version.