ci: release tag_name to include 3 crate versions
Rather than just having the `safe_network` crate version, the tag name field will now include the
version numbers of the 3 crates. This will enable us to be able to install a specific version of the
CLI. This tag name should always be unique, since there will always be at least 1 crate with a
change in it.

The title of the release will also include the names and versions of each of the components, even if
one crate didn't change.

The commit message will still just include the crates that changed, so that we know which publishing
stages to run.

Also fixes a little bug here with one of the asset hashes not being set correctly.