chore: remove `self-update` feature from `node install`
This feature should really never have been applied to `node install`, since self update for the node
occurs on the node binary itself. We only want to disable self update for the CLI and we still want
the `node install` feature.

This means that the `self-update` feature itself isn't actually tied to the `self_update` crate,
which will still be used, because parts of its functionality are used for the `node install`
command.

It will now just disable the CLI itself from updating, which a separate concern from the `node`
command.

This also fixes a little mistake with the tag_name for the release not being assigned correctly.