ci: upload artifacts to s3 buckets
We removed this in favour of using Github Releases as the primary mechanism for downloading the
binaries, but this is an issue when everything is in the same repository and each release has
binaries for both `safe` and `sn_node`. It means we can't easily retrieve specific versions of any
individually component, which we may want to do to give users the ability to install specific
versions of the binaries.

For this reason, we'll also upload the binaries to S3. So when the user requests a specific version,
we'll go to S3 to fetch it, but we can get the latest version from Github Releases.

There's also a little refactor here to use the `gh` command line tool to upload the artifacts as
assets to the release. Due to the fact that we have lots of artifacts, using the action is a very
verbose way of doing the upload.