refactor: use s3 as download source for sn_node
Now that we've restructured the release process such that one release has all crates, it makes it
more difficult and awkward to install specific versions of any of the binaries, when Github Releases
are the source of the download. This is because you can't query Github releases to search for the
version number of one component. If you want to get a particular version of something by a
convention (like say the tag name on the release), you need to know the version number for each of
the 3 components.

Here the `node install` command is changed to use the Github API to get the latest version, but once
the version is obtained, we download the binary from S3. This is so we can also support downloading
specific versions. Using the S3 bucket, we can do this by convention, based on the version number
and the target platform. It happens to be the case that the self update crate provides some
functions for querying the API and also downloading and extracting a binary from a compressed tar
archive, so we're still making use of that crate here.