chore: re-introduce version arg for cli install script
The release process has been modified to upload all our releases to S3, so we can re-introduce the
`--version` argument for installing specific versions. If the user supplies a version, we can get
that from S3. Otherwise, we'll get the latest version from Github Releases.

Also run a `cargo fmt` after a rebase with some conflicts.