refactor(elder targets): rename + use glob const
- Adds max_num_faulty_elders based on const assumption for
max number of faulty Elders as 1/3.
- Adds at_least_one_correct_elder based on max_num_faulty_elders.

These numbers were made dependent on global consts, as was asked for in
the todo comment prior to this change.

Also, the number of Elders to whom a client sends a chunk is not based
on the chunk copy count, but on our assumptions on max number of faulty
Elders, hence renaming the variable.