ci: version bumping process to use 3 crates
Previously we had an issue where `smart-release` would error if `sn_api` or `sn_cli` had a change,
but `safe_network` didn't. If it's invoked using all 3 crates, this doesn't happen. I was sure I had
tried this mechanism before and it didn't work, but that may have been with an earlier version.

I also instructed the script to exit with an error when the `smart-release` output contains
'Manifest version of provided package' in its output. This is intended to detect such output as:
```
[INFO ] Manifest version of provided package 'sn_cli' at 0.43.1 is sufficient to succeed latest
released version 0.43.0, ignoring computed version 0.43.2
```

What this means is, a crate had its version bumped, a tag was pushed and created, but it failed to
publish. One such scenario is when separate API and CLI PRs are merged in quick succession. The CLI
might reach publishing first, but the version of the API it expects to exist hasn't been published
yet.