chore: dont fail client send w/ closed conn
Closing a connection would cause a client error in send before this.
Now we check if we had an application layer close... something intentional.
If so we do not consider this a failure at the client and can
continue thereafter.

qp2p is upgraded to send a reason on close that we can check against.