ci: switch off verbose logging for cli tests
The CLI tests parse output from the CLI and they expect it to be in a certain form, so any
additional logging needs to be switched off.

The other tests in the PR workflow switch verbose logging on because it's useful for those tests and
there are also some scripts that parse the output to try and determine if the network is ready.