feat: retry intial client contact if it fails
Previously we tried to connect to all known nodes, but if we reached the end of the list, we'd keep hitting the same nodes over and over.
Now we fail after trying all candidates, and retry with a new xorname to get new candidates (if we know a prefixmap).
We also mark the attempted connections as failed, so the client attempts to create
fresh new connections instead of using possibly dead connections