refactor: nrs map fetch and rename multimap
The `MultimapKeyValues` type alias was renamed to `Multimap` to make it easier to work with. Some
functions were also renamed to the same effect, e.g., `fetch_multimap_values` -> `fetch_multimap`.

The function for retrieving the NRS map was also refactored. Some aspects were broken down into
smaller functions to reduce some complex if/else cases. Things were also setup for the version in
the last entry in the Multimap to be collected. This is going to be used with the returned
NrsMapContainer, to facilitate returning links to immutable content, which doesn't have a version.
Currently the code isn't setup to support that scenario.