feat: retrieve immutable content via nrs
Changes the resolver to enable retrieval of NRS links to immutable content. Previously, the
NRS `target_url` was incorrectly being checked to see if it contained a version. We now validate
this URL using the same process used by NRS, which won't assert that a file link has a version. To
reduce duplication, the `validate_nrs_url` function in the NRS module was changed to public.

Tests were added to both the API and CLI to cover the scenario.

The NrsMap struct was extended to include a subname_version field. This is going to be used to
request the map with a particular version of a subname. If no version was specified when the
container was retrieved, then the field won't be set. This is why it's an Option. Since we're
storing this field on the NrsMap, the `version` field on the `NrsMapContainer` was removed.

A couple of CLI NRS tests were also re-enabled, one of which happened to be related to immutable
content.