ci: remove smart release dry run
Previously, the dry run was being used to detect a string of text I had seen appear when version
bumping succeeded but publishing of one or more crates failed.

Two days ago, Smart Release had a new release (v0.8.0) and it seems this assumption no longer holds.
The text is appearing now, but the version it's referring to has had its crate published. This was
the only thing the dry run was checking, so it doesn't serve any purpose now.

The scheduled time for the nightly run was also pushed back a couple of hours, as we'd seen Bors
still running and pushing to main after 0230.