ci: manual trigger for nightly run
Uses the `workflow_dispatch` event to enable manual triggering of the nightly run, which can be
useful if we want to release during the day.

Also reintroduces the dry run for Smart Release. I mistakenly believed this was only being used to
check for problems with publishing, forgetting it was also used to detect which crates have changes.
That info is used in the release commit message to specify the crates with changes, which is
subsequently used by the release workflow so that publishing only runs for crates with version bumps.