chore(sn/node): only prioritise service cmd/query with permits.
Previously we were holding permits in a RwLock struct for _any_ command.

Now we remove both the conecpt of permits here, removing one potential
locking loction, and the unused message prioritisation is removed to
clean that up.

Now we simply check if we have max allowed pending queries at any one
time, and if so, we drop inbound queries.