feat: store full public name in nrs map
Entries will now use the full public name reference, rather than the just subname. Like so:
* example -> link
* a.example -> link
* b.example -> link
* a.b.example -> link

This is to facilitate resolving NrsMapContainer content using its XorUrl, which effectively prints
the NRS map. It was a feature we had that broke at some point. Right now, the XorUrl of the
container will resolve the content the topname is linked to, or will result in an error if the
there's no topname link.