refactor: remove url sanitisation from api
The `Safe::parse_url` function was removed from the API. This function 'sanitised' a URL by applying
a `safe://` prefix to a URL string if the caller hadn't specified it.

Initially, it was done to tidy up NRS code that was calling this function, but the same code was
also calling a private function `parse_url` was making use of, so effectively the code was being
called twice. More generally, we decided callers of the API should be responsible for passing a
valid URL.

The function was being called by various other parts of the API and also in the CLI, so these were
changed to call `SafeUrl::from_url` directly.

Some code was added to CLI commands to apply the `safe://` prefix if the user omitted it, so no
functionality  was broken. A few test cases were added to cover it. A couple of NRS test cases for
validating URLs were also removed as they no longer apply. This behaviour may actually have been
incorrect in the first place.

Also apply various clippy fixes.