tests: update nrs tests to use helper
The helper was moved into the test_helpers module and renamed to TestDataFilesContainer, since this
seems more accurate.

A test was added for retrieving a topname link. This is arguably covered by the `nrs_associate`
tests, but I think it's good to have this test to call it out as functionality we don't want to
break.