chore: simplify ae msg handling at clients
Previously AE-retry/redirect did not both have a SAP. Now
they do we can simplify that and directly use it in either case.

We also can remove the AE caches, and just send on each incoming AE-msg
to one elder (as opposed to `count` elders). This is deterministic based
on sender name. Which means we can simply do it as each one comes in. No
tracking of prior messages needed.

We also remove the `make_contact` exemptions in case this interferes
with initial network contact.

Overall simpler and cleaner.