feat: resolve nrs map container content
The resolver can now return `NrsMapContainer` content, which can then be displayed by the CLI with
the `cat` and `dog` commands. This functionality was unintentionally broken at some point.

The first change introduced an `NrsEntry` field in `SafeData`, and modified the `NrsMapContainer` to
remove its `resolve_into` and `public_name` fields. The intention is for the resolver to return
`NrsMapContainer` data when a container XOR-URL is used, but when using an NRS URL, an `NrsEntry`
will be returned. The `NrsMapContainer` data will have the NRS map, whereas the `NrsEntry` will only
contain the target link and subname version. It's worth noting, the `NrsEntry` doesn't have an
XOR-URL because the entries are still stored in the map. An NRS URL still has an `NrsMapContainer`
content type and that content is retrieved during the resolution process.

This brings us to the next change. The `nrs_get` API was modified to return an `Option<SafeUrl>`,
where `None` will now be returned if the container XOR-URL is used. In this case, the resolver will
know to return `NrsMapContainer` data, otherwise, it will return the `NrsEntry` with the target URL.
One exception is worth mentioning: if the NRS URL uses the registered topname and that topname
*doesn't* link to anything, `NrsMapContainer` data will also be returned. To make these extensions,
small unit tests were added to the `NrsMap` and several tests were added to the resolver to cover
these scenarios.

With these changes in place, the CLI could then be updated. The `cat` and `dog` commands were
modified to print the NRS map when `NrsMapContainer` data was returned. Previously, the map was
printed as a table, but this isn't really suitable for presentation because the table crate doesn't
have the ability to use multi-line cells and the target links are too large, so I changed it to
print a list. Test cases were added for both commands, which should hopefully prevent us breaking
the feature again.

Finally, some usability changes were also made to `nrs` commands to give the user the XOR-URL of the
container. This can be useful to them if they want to list all the entries in a map.