chore: sudo based installation for cli
Give users the option to run the CLI installer as sudo.

In this case, the CLI will be installed to /usr/local/bin, which is very common for installations on
Linux. That location is almost always already on PATH, even in minimal setups. This removes the need
for modifying any of the user's shell configurations, which some users don't like. I've also
personally dealt with users who had a problem when the installer wasn't modifying their shell
configuration correctly.

Some simple installer tests were added to the nightly configuration and the README was also updated
with new instructions, which included fixing an invalid link to the install script.