chore: powershell install script for cli
Provide a Powershell install script for Windows users. This removes the requirement for using Git
Bash or other Linux emulation to setup the CLI.

The install process is:
* Download and install the Visual C++ Redistributable
* Create the `%USERPROFILE%\.safe\cli` directory
* Add `%USERPROFILE%\.safe\cli` to the user `PATH` variable
* Download and unpack the latest version of `safe` to `%USERPROFILE%\.safe\cli`

A few points worth noting. First, the script can run without admin access, but the VC++ install
requires that, so it will elevate to admin to run that. Second, the script is idempotent, so it will
run OK, if, e.g., VC++ is already installed. Third, the addition of the `PATH` variable
unfortunately requires users to start a new Powershell session. Finally, if VC++ was installed, that
requires a machine restart. The script informs users of either of those, and they are only required
the first time.

I tested it on a minimal Windows installation on a VM, before the VC++ prerequisite had been
installed. It worked as intended in this setup. A test was added to the nightly run.

The README was updated to provide the user with instructions for running the script.