docs: further updates for cli user guide
Several changes were made to bring the user guide up-to-date:
* Move the XorUrl section to near the top of the document, below the network setup section. I done
  this because I think it makes sense to explain the URL concept early, since it's used by every
  command.
* Add a new section on Keys, where it's relevant to current network functionality. The previous Keys
  documentation was removed because it referred to out-of-date features.
* Re-write the NRS, Cat and Dog documentation to make it a bit leaner.
* Remove the section on shell completions. I'm going to come back to this very shortly and possibly
  include the completions to install them along with the application.
* Remove links to the 'Safe Dev Hub' from the Further Help section.

The new sections have been written with a line length of 100. This is just to make it easier to read
while it's being written, without having to use word wrapping, which mangles the tables that are
output by some CLI commands. Markdown will still render the document with a larger line length,
which I think will wrap based on the size of the browser window.