ci: use nextest as test runner
The PR tests have been changed to use Nextest to perform the test runs. Nextest can produce test
reports in the popular Junit format, which means we can publish the test results along with the PR.
Since it was tedious to copy and upload the reports as artifacts, this was encapsulated in a custom
`cargo-nextest` action.

To support publishing test results from forks, the publish occurs in another workflow. See the
documentation:
https://github.com/EnricoMi/publish-unit-test-result-action/blob/v1.20/README.md#support-fork-repositories-and-dependabot-branches

The following changes were made to the existing run:
* Remove the `build` job. Everything performed in this job is also now performed in the test stages,
  so as far as I can see, it doesn't serve a purpose any more. Since it was running concurrently,
  removing it won't save any time, but it was now using unnecessary resources.
* Remove use of the `paths-filter` action. The unit tests are quick to run, so we may as well just
  run all of them.
* The client tests are performed in a single run rather than being split into multiple runs.
  Previously this was setup to make it easier to identify failures, but these tests seem to be quite
  stable now.
* Remove the network asserts tests. These were a small set of tests that always failed, but were set
  to continue on error.
* Run the CLI unit tests.
* Remove the CLI integration tests. They don't run stable when running against a local network on
  the Github build agent.

Despite the CLI integration tests being removed, a fix went in for the run to be performed with
multiple threads.

Previously, the CLI test run was setup to use `--test-threads 1`. This was because, if we ran with
multiple threads, sometimes we'd see an error occur when trying to read the networking config file.
When multiple threads are used, this means you're running multiple instances of `safe` concurrently,
since the tests are integration tests that run the binary. Somehow, a race condition was occurring,
where one instance reads the config file, while another writes to it and causes it to be empty. The
instance that reads the file expects it to be populated, but the other has cleared it, and an error
occurs when trying to deserialize an empty byte array. I examined the tests and couldn't find out
how it was possible for the file to be empty, but I added some code to accommodate the scenario
anyway. This means we can run the test suite using multiple threads, which significantly decreases
the running time.