ci: use nextest for nightly runs
Like CI, the nightly release run has now been updated to use Nextest.

Each of the three test runs are now performed in parallel and we're going to attempt to use ten
threads for the CLI run.

The network we perform the tests against has been updated to use 15 nodes.