ci: run cli tests after api and client
Running all three jobs at the same time didn't seem to be stable.

Going to try now with the client and API tests running at the same time, then let the CLI tests run
on their own, but still try with 10 threads.

Also added an explicit build step to the client tests, to avoid timeouts when performing the test
run.