ci: re-enable cli tests for pr tests
Foolishly, I had forgotten to include key generation before the test run, which hopefully explains
the failures I was seeing in the previous runs.

This has also been included for the nightly run too.

With this done, the CLI tests should be able to run for PRs, as it was before.