feat!: upgrade to new version of qjsonrpc
The `sn_api` crate is updated to use the new version of `qjsonrpc`.

The `qjsonrpc` library had breaking changes to pass cert and key paths rather than passing a
directory that was assumed to contain them. This change caused `sn_api` to do the same. There is an
unfortunate consequence here: to use `SafeAuthdClient` you need to pass a certificate, along with
another certificate and a private key for starting an endpoint that deals with notifications for
approving authentication requests. The previous setup was just making the assumption that both keys
and certificates were in the same directory. Since this change led to an odd interface for
`SafeAuthdClient` we had some discussion around using a different mechanism for notifications, but
we decided to come back to it later with a wider review.

The API change forced an update to some code in the CLI, even though the authd system isn't really
being used at the moment. The self-signed certificates are now being generated in the CLI. Since the
CLI is a specific application, I think this is a more appropriate place for that to happen.