fix(sn): Improve query handling, keep peers on DataNotFound
Previously if we had DataNotFound, any waiting peers are not requeued.
So this means that one bogus adult can effectively fail a query which
might otherwise succeed.

In doing so, we need to always resend queries to adults if clients retry
as otherwise they may remain queued but wont actually requery adults
ever again.

This appears to improve test time, even after reducing the RETRY_COUNT
on messages.