feat: optimise DataQuery cache
Move the pending_requests to storing an Arc<DashSet<Peers>>

This is due to cache providing a clone when we , which means
we may not always be setting pending peers correctly.

Now we also use the DATA_QUERY_TIMEOUT, set to 15s to decide if we should
re-send a query. (Previously it was only sent if no peer was waiting; ).

This means we have a hard limit on how often we may ask for a chunk if peers are waiting for
that chunk.