fix: node install command to use correct version
With the addition on the dysfunction crate, this introduced a new version number into the tag name.
The parsing had to be updated for this.

I've also re-enabled the test that would have caught this. It shouldn't be as much of a problem with
nextest, which has the ability to retry tests. I've seen the rate limited stuff passing ok on
retries. If we see it fail, we can disable it again.