feat: compare against all nodes in section
When calculating scores, we compare against the average of all the nodes in the section, rather than
the 'neighbours' of the node. As a consequence, the `DysfunctionDetection` struct becomes simpler,
as we don't need to keep track of what nodes are 'closer' to others as the set of nodes changes.

In the course of this change, the Dysfunction API was updated.

First, the `ScoreType` enum was renamed `IssueType`, which now details the three issue types, rather
than have timed versus operational. Secondly, one `track_issue` function merges three
`track_x_issue` functions for adding issues to the tracker.

Both these decisions were influenced by property testing: there were properties that should have
been true for each issue type, and it was easier to write tests for those properties if they all
used a single API.

It's worth noting, the `track_issue` function provides an `Option` argument for supplying an
`op_id`. This value only applies when a `PendingRequestOperation` type is used. At first, the
pending request entry was declared as `PendingRequestOperation([u8; 32])`, which makes sense
initially. For adding the issue, you just pass the op ID along with the issue type. However, the
problem comes when you later want to select the issues of this type. For example:
```
let _ = op_scores.insert(
    *node,
    self.calculate_node_score(
        node,
        adults.clone(),
        IssueType::PendingRequestOperation,
    )
    .await,
);
```

If you have the `op_id` parameter on the enum entry, the code becomes very clunky:
```
let _ = op_scores.insert(
    *node,
    self.calculate_node_score(
        node,
        adults.clone(),
        IssueType::PendingRequestOperation([1; 32])
    )
    .await,
);
```

As can be seen, you need to supply an ugly placeholder value that has no effect. For this reason, I
decided to just supply the `op_id` as an optional argument on `track_issue`. Neither solution is
completely ideal, but I prefer this one.

Remaining changes were a bit more superficial. A `ScoreResults` helper struct was introduced to
return all three types of scores from the refactored `calculate_scores` function, rather than
returning a tuple of three hash tables, which would have been a bit cumbersome.